feat: UI 구성

+++ careerForDeveloper/src/main/webapp/reactjs/src/components/DropUser.js
@@ -1,17 +1,19 @@
 import React from 'react';
 import {useNavigate} from 'react-router-dom';
 import axios from "axios";
-import {useSelector} from "react-redux";
+import {useDispatch, useSelector} from "react-redux";
+import {DELETE_TOKEN, SET_TOKEN} from "../Store/Auth";
+import {getCookieToken, removeCookieToken} from "../utils/Cookie";
 
 function DropUser() {
     const [pwd, setPwd] = React.useState("");
     const navigate = useNavigate();
+    const dispatch = useDispatch();
     const token = useSelector(state => state.authToken);
 
     console.log(token);
 
     const dropUser = (e) => {
-        console.log(token);
         if (!pwd) {
             return alert("비밀번호를 입력하세요");
         }
@@ -21,21 +23,37 @@ function DropUser() {
         const res = confirm("정말 탈퇴하시겠습니까?");
 
         if (res) {
-            axios.delete("/users", {
-                data: {
-                    pwd: pwd
+            const refreshToken = getCookieToken();
+            if (!refreshToken) return;
+            axios.get("/auth", {
+                headers: {
+                    "REFRESH-TOKEN": refreshToken,
                 }
+            }).then((res) => {
+                console.log(res.data);
+                if (!res.data.isSuccess) {
+                    return alert(res.data.message);
+                }
+                axios.defaults.headers.common['X-ACCESS-TOKEN'] = res.data.result.accessToken;
+                dispatch(SET_TOKEN(res.data.result.accessToken));
+                axios.delete("/users", {
+                    data: {
+                        pwd: pwd
+                    }
+                })
+                    .then((res) => {
+                        console.log(res);
+                        if (res.data.isSuccess) {
+                            alert("성공적으로 탈퇴했습니다.");
+                            dispatch(DELETE_TOKEN());
+                            removeCookieToken();
+                            navigate("/");
+                        } else {
+                            alert(res.data.message);
+                            navigate(-1);
+                        }
+                    });
             })
-                .then((res) => {
-                    console.log(res);
-                    if (res.data.isSuccess) {
-                        alert("성공적으로 탈퇴했습니다.");
-                        navigate("/");
-                    }
-                    else {
-                        alert(res.data.message);
-                    }
-                });
         }
     };
